L4G-8 Update test package and class name [RDZ, FC]
