+++ Client/app/src/main/java/com/Client/Table/ui/bio/EditBioFragment.kt
@@ -4,12 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.Client.Table.R
-import com.Client.Table.ui.slideshow.SlideshowViewModel
 
 class EditBioFragment : Fragment() {
